% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JefferysBioStatTools-package.r
\docType{package}
\name{JefferysBioStatTools}
\alias{JefferysBioStatTools}
\alias{JefferysBioStatTools-package}
\title{JefferysBioStatTools - Biostatistic tools organized for easy use}
\description{
Statistical tools are scattered all over the place in R. I want one place
to find them. I also want \emph{my} documentation for them to go somewhere
directly accessible, not hidden in text files. I expect most things here
will be wrappers around existing functions. It does add overhead to pass big
data, have to be careful about that.
}
